requirements.txt
line-bot-sdk
fastapi
uvicorn